Buzzer: invert meaning of pause (0xF) and end (0x0)
I have noticed that sometimes (due to human error, usually) there's a missing
end note and buzzer goes on and on playing pauses (because uninitialized memory
is read as 0). So it seems appropriate using 0x0 as end note and 0xF as pause,
so if, by chance, an end note is missing, the buzzer will try to access
uninitialized memory and read that as an end note. This of course fails if there
is used (and thus initialized) memory contiguous to the location of the notes
buffer, but we can't do everything, can we?

So, in short, pause is now 0xF and end is 0x0. All ringtones already in the code
have been changed and rtttl2bin script has been updated.