Initial asynchronous buzzer implementation
This changeset implements the buzzer in a non-blocking way. There are a few
gotchas though:
- Buzzer uses SMCLK for tone generation, so the deepest LPM mode we can go into
is LPM1
- Since we don't have much control over the notes array passed to buzzer_play,
it copies the array into a malloc-ed buffer. This of course means we need to
know the length of the array, which we need to find out by traversing it. That
is admittedly inefficient.

Additionally, a system bus message is sent when the buzzer finishes ringing.

A courtesy inline function (sleep_and_service_wd) was also create to keep the
main loop clean.